Changed fetchType to i16

+++ impala/thrift/TCLIService.thrift
@@ -1057,7 +1057,7 @@ struct TFetchResultsReq {
   3: required i64 maxRows
 
   // Specify the fetchType for retrieval of logs use 1
-  4: optional i32 fetchType
+  4: optional i16 fetchType
 }
 
 struct TFetchResultsResp {
